backend/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
protobuf==3.20.1
requests==2.32.3
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
typing_extensions==4.12.2
